added hash_many functions, moved missplaced tests
