+++ src/commitment_scheme/merkle_tree.rs
@@ -62,3 +62,23 @@ impl<T: Hasher> MerkleTree<T> {
         res
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::commitment_scheme::{blake3_hash::Blake3Hasher, merkle_tree::MerkleTree};
+
+    #[test]
+    fn merkle_tree_building() {
+        let leaves = [0; 64];
+        let mut tree: MerkleTree<Blake3Hasher> = MerkleTree::commit(&leaves[..]);
+        assert_eq!(tree.height, 3);
+        assert_eq!(
+            tree.data[0][0].to_string(),
+            "4d006976636a8696d909a630a4081aad4d7c50f81afdee04020bf05086ab6a55"
+        );
+        assert_eq!(
+            tree.root_hex(),
+            "06253c52ed8536e4b07757d679c547fdb2051181a9cbd1e3516bfc71742936f7"
+        )
+    }
+}
